Refactor the search form into its own component
